requirements.txt
pickle5==0.0.13
waitress==2.1.2
python-dotenv==1.0.1
absl-py>=2.1.0
astunparse>=1.6.3
audioread>=3.0.1
blinker>=1.8.2
gunicorn==20.1.0
certifi>=2024.2.2
cffi>=1.17.1
Flask-Cors>=4.0.0
charset-normalizer>=3.3.2
click>=8.1.7
colorama>=0.4.6
decorator>=5.1.1
Flask>=3.0.3
flatbuffers>=24.3.25
gast>=0.6.0
google-pasta>=0.2.0
grpcio>=1.66.2
h5py>=3.12.1
idna>=3.7
itsdangerous>=2.2.0
Jinja2>=3.1.4
joblib>=1.4.2
keras>=3.5.0
lazy_loader>=0.4
librosa>=0.10.2.post1
llvmlite>=0.43.0
Markdown>=3.7
matplotlib>=3.9.2
numpy>=1.26.4
opencv-python>=4.10.0.84
pandas>=2.2.2
protobuf>=4.25.5
requests>=2.31.0
rich>=13.7.1
scikit-learn>=1.5.1
scipy>=1.14.0
soundfile>=0.12.1
soxr>=0.5.0.post1
TensorBoard>=2.17.0
tensorflow>=2.17.0
Werkzeug>=3.0.3
